+++ app/src/main/java/site/weshare/android/presentation/gonggu/RegisterDetailScreen.kt
@@ -1,13 +1,10 @@
 package site.weshare.android.presentation.gonggu
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.verticalScroll
@@ -17,10 +14,8 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
@@ -28,12 +23,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import androidx.compose.ui.zIndex
 import site.weshare.android.R
-import java.text.SimpleDateFormat
-import java.util.*
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -415,421 +405,6 @@ fun RegisterDetailScreen() {
 
             Spacer(modifier = Modifier.height(16.dp))
         }
-
-        if (showCalendarModal) {
-            CalendarModal(
-                selectedDate = selectedDate,
-                onDateSelected = { date ->
-                    selectedDate = date
-                    showCalendarModal = false
-                },
-                onDismiss = { showCalendarModal = false }
-            )
-        }
-
-        if (showPriceInfoModal) {
-            PriceInfoModal(onDismiss = { showPriceInfoModal = false })
-        }
-
-        // 정보 모달
-        if (showInfoModal) {
-            InfoModal(
-                onDismiss = { showInfoModal = false },
-                guestCount = guestCount,
-                myGuestCount = myGuestCount
-            )
-        }
-    }
-}
-
-@Composable
-fun CalendarModal(
-    selectedDate: String,
-    onDateSelected: (String) -> Unit,
-    onDismiss: () -> Unit
-) {
-    // 현재 날짜 파싱
-    val dateFormat = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault())
-    val currentDate = try {
-        dateFormat.parse(selectedDate) ?: Date()
-    } catch (e: Exception) {
-        Date()
-    }
-
-    val calendar = Calendar.getInstance().apply { time = currentDate }
-    var currentYear by remember { mutableStateOf(calendar.get(Calendar.YEAR)) }
-    var currentMonth by remember { mutableStateOf(calendar.get(Calendar.MONTH)) }
-    var tempSelectedDate by remember { mutableStateOf(selectedDate) }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.5f))
-            .clickable { onDismiss() },
-        contentAlignment = Alignment.Center
-    ) {
-        // 통합 카드 (상하단만 둥근 모서리)
-        Card(
-            shape = RoundedCornerShape(12.dp),
-            elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-            colors = CardDefaults.cardColors(containerColor = Color.White),
-            modifier = Modifier
-                .padding(24.dp)
-                .offset(x = (5).dp, y = 200.dp)
-                .clickable(enabled = false) { }
-        ) {
-            Column {
-                // 달력 영역 (상단 둥근 모서리)
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .background(
-                            Color(0xFFD6F0C7),
-                            RoundedCornerShape(topStart = 12.dp, topEnd = 12.dp)
-                        )
-                ) {
-                    Column(
-                        modifier = Modifier.padding(16.dp)
-                    ) {
-                        // 헤더 - 년월 네비게이션
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.Center,
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Icon(
-                                Icons.Default.KeyboardArrowLeft,
-                                contentDescription = "이전 달",
-                                modifier = Modifier
-                                    .size(32.dp)
-                                    .clickable {
-                                        if (currentMonth == 0) {
-                                            currentMonth = 11
-                                            currentYear--
-                                        } else {
-                                            currentMonth--
-                                        }
-                                    }
-                            )
-
-                            Text(
-                                text = "${currentYear}년",
-                                fontSize = 18.sp,
-                                fontWeight = FontWeight.ExtraBold,
-                                modifier = Modifier.padding(horizontal = 13.dp)
-                            )
-
-                            Text(
-                                text = "${currentMonth + 1}월",
-                                fontSize = 18.sp,
-                                fontWeight = FontWeight.ExtraBold,
-                                modifier = Modifier.padding(end = 13.dp)
-                            )
-
-                            Icon(
-                                Icons.Default.KeyboardArrowRight,
-                                contentDescription = "다음 달",
-                                modifier = Modifier
-                                    .size(32.dp)
-                                    .clickable {
-                                        if (currentMonth == 11) {
-                                            currentMonth = 0
-                                            currentYear++
-                                        } else {
-                                            currentMonth++
-                                        }
-                                    }
-                            )
-                        }
-
-                        Spacer(modifier = Modifier.height(16.dp))
-
-                        // 요일 헤더
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceEvenly
-                        ) {
-                            listOf("일", "월", "화", "수", "목", "금", "토").forEach { day ->
-                                Text(
-                                    text = day,
-                                    fontSize = 14.sp,
-                                    fontWeight = FontWeight.Medium,
-                                    color = Color.Black,
-                                    textAlign = TextAlign.Center,
-                                    modifier = Modifier.width(40.dp)
-                                )
-                            }
-                        }
-
-                        Spacer(modifier = Modifier.height(8.dp))
-
-                        // 달력 그리드 생성
-                        val calendarDays = generateCalendarDays(currentYear, currentMonth)
-
-                        calendarDays.chunked(7).forEach { week ->
-                            Row(
-                                modifier = Modifier.fillMaxWidth(),
-                                horizontalArrangement = Arrangement.SpaceEvenly
-                            ) {
-                                week.forEach { dayInfo ->
-                                    val (day, isCurrentMonth) = dayInfo
-                                    val dateString = String.format("%04d-%02d-%02d", currentYear, currentMonth + 1, day)
-                                    val isSelected = isCurrentMonth && tempSelectedDate == dateString
-
-                                    Box(
-                                        modifier = Modifier
-                                            .size(40.dp)
-                                            .clickable(enabled = isCurrentMonth) {
-                                                if (isCurrentMonth) {
-                                                    tempSelectedDate = dateString
-                                                }
-                                            }
-                                            .background(
-                                                if (isSelected) Color(0xFF2FB475) else Color.Transparent,
-                                                CircleShape
-                                            ),
-                                        contentAlignment = Alignment.Center
-                                    ) {
-                                        Text(
-                                            text = if (day > 0) day.toString() else "",
-                                            fontSize = 14.sp,
-                                            color = when {
-                                                isSelected -> Color.White
-                                                isCurrentMonth -> Color.Black
-                                                else -> Color.Gray
-                                            },
-                                            fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Normal
-                                        )
-                                    }
-                                }
-                            }
-                            Spacer(modifier = Modifier.height(4.dp))
-                        }
-                    }
-                }
-
-                // 구분선 (선택사항)
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(1.dp)
-                        .background(Color(0xFFE0E0E0))
-                )
-
-                // 버튼 영역 (하단 둥근 모서리)
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .background(
-                            Color.White,
-                            RoundedCornerShape(bottomStart = 12.dp, bottomEnd = 12.dp)
-                        )
-                ) {
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        horizontalArrangement = Arrangement.SpaceEvenly
-                    ) {
-                        Button(
-                            onClick = { onDismiss() },
-                            colors = ButtonDefaults.buttonColors(
-                                containerColor = Color(0xFF73CDA2)
-                            ),
-                            shape = RoundedCornerShape(8.dp),
-                            modifier = Modifier
-                                .weight(1f)
-                                .height(48.dp),
-                            contentPadding = PaddingValues(horizontal = 16.dp)
-                        ) {
-                            Text(
-                                text = "취  소",
-                                color = Color.White,
-                                fontSize = 16.sp,
-                                fontWeight = FontWeight.Medium
-                            )
-                        }
-
-                        Spacer(modifier = Modifier.width(16.dp))
-
-                        Button(
-                            onClick = {
-                                onDateSelected(tempSelectedDate)
-                            },
-                            colors = ButtonDefaults.buttonColors(
-                                containerColor = Color(0xFF2FB475)
-                            ),
-                            shape = RoundedCornerShape(8.dp),
-                            modifier = Modifier
-                                .weight(1f)
-                                .height(48.dp),
-                            contentPadding = PaddingValues(horizontal = 16.dp)
-                        ) {
-                            Text(
-                                text = "선  택",
-                                color = Color.White,
-                                fontSize = 16.sp,
-                                fontWeight = FontWeight.Medium
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-
-@Composable
-fun PriceInfoModal(
-    onDismiss: () -> Unit
-) {
-    // 투명 배경 + 클릭 시 닫기
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.5f))
-            .clickable { onDismiss() }, contentAlignment = Alignment.Center
-    ) {
-        // 팝업 카드
-        Card(
-            shape = RoundedCornerShape(8.dp),
-            elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-            modifier = Modifier
-                .wrapContentSize()
-                .align(Alignment.TopEnd)    // 가격 아이콘 바로 위에 띄우려면 위치 조정
-                .offset(x = (-16).dp, y = 620.dp) // 필요에 따라 x,y 조정
-                .size(width = 360.dp, height = 125.dp)
-        ) {
-            Column(modifier = Modifier.padding(16.dp)) {
-                // 상단: info 아이콘 + 제목 + 닫기 버튼
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.info),
-                        contentDescription = null,
-                        tint = Color.Black,
-                        modifier = Modifier.size(20.dp)
-                    )
-
-                    Icon(
-                        painter = painterResource(id = R.drawable.close),
-                        contentDescription = "닫기",
-                        tint = Color.Black,
-                        modifier = Modifier
-                            .size(17.dp)
-                            .clickable { onDismiss() }
-                    )
-                }
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 8.dp)
-                        .horizontalScroll(rememberScrollState()),  // ★화면에 다 안 들어올 때 스크롤
-                    horizontalArrangement = Arrangement.Start,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = "등록 상품이 무료 배송 상품이예요 →",
-                        fontWeight = FontWeight.Bold,
-                        letterSpacing = (-0.8).sp,
-                        fontSize = 15.sp,
-                        // weight 제거!
-                    )
-                    Spacer(modifier = Modifier.width(4.dp))
-                    Image(
-                        painter = painterResource(id = R.drawable.freedelivery),  // PNG 혹은 벡터 원본
-                        contentDescription = "무료 배송(모형)",
-                        modifier = Modifier
-                            .height(30.dp)
-                            .wrapContentWidth()
-                            .alignByBaseline()  // 텍스트 기준선 맞춤
-                    )
-                }
-
-                // 예시 문구
-                Text(
-                    text = "등록 상품의 배송비가 있어요 → 배송비를 입력해주세요",
-                    fontWeight = FontWeight.Bold,
-                    letterSpacing = (-0.8).sp,
-                    fontSize = 15.sp,
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun InfoModal(
-    onDismiss: () -> Unit,
-    guestCount: Int,
-    myGuestCount: Int
-) {
-    // 반투명 배경 + 클릭시 닫기
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Black.copy(alpha = 0.5f))
-            .clickable { onDismiss() }
-    ) {
-        // 가운데 박스
-        Card(
-            shape = RoundedCornerShape(12.dp),
-            elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-            modifier = Modifier
-                .align(Alignment.Center)
-                .size(width = 290.dp, height = 180.dp)
-                .padding(24.dp)
-                .offset(x = 39.dp, y = 150.dp)
-                .wrapContentWidth()
-        ) {
-            Column(modifier = Modifier.padding(16.dp)) {
-                // 상단: info 아이콘 + 제목 + 닫기 버튼
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.info),
-                        contentDescription = null,
-                        tint = Color.Black,
-                        modifier = Modifier.size(20.dp)
-                    )
-
-                    Icon(
-                        painter = painterResource(id = R.drawable.close),
-                        contentDescription = "닫기",
-                        tint = Color.Black,
-                        modifier = Modifier
-                            .size(17.dp)
-                            .clickable { onDismiss() }
-                    )
-                }
-
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = "공동 구매에서 내가 가져갈 개수를 선택해주세요",
-                    fontWeight = FontWeight.Bold,
-                    letterSpacing = (-0.8).sp,
-                    fontSize = 15.sp,
-                    modifier = Modifier.fillMaxWidth()
-                )
-                // 예시 문구
-                Text(
-                    text = "예) 총 구매 개수 100, 내가 가져가고 싶은 개수 10 \n \t\t\t→ 10개를 선택해주세요",
-                    fontSize = 11.sp,
-                    letterSpacing = (-0.8).sp,
-                    lineHeight = 13.sp,
-                    color = Color.DarkGray
-                )
-            }
-        }
     }
 }
 
@@ -996,76 +571,10 @@ fun CounterRow(
     }
 }
 
-fun generateCalendarDays(year: Int, month: Int): List<Pair<Int, Boolean>> {
-    val calendar = Calendar.getInstance()
-    calendar.set(year, month, 1)
-
-    // 해당 월의 첫 날이 무슨 요일인지 (일요일=1, 월요일=2, ...)
-    val firstDayOfWeek = calendar.get(Calendar.DAY_OF_WEEK) - 1
-
-    // 해당 월의 마지막 날
-    val lastDayOfMonth = calendar.getActualMaximum(Calendar.DAY_OF_MONTH)
-
-    // 이전 달의 마지막 날
-    calendar.add(Calendar.MONTH, -1)
-    val lastDayOfPrevMonth = calendar.getActualMaximum(Calendar.DAY_OF_MONTH)
-
-    val days = mutableListOf<Pair<Int, Boolean>>()
-
-    // 이전 달의 날짜들 (회색으로 표시)
-    for (i in firstDayOfWeek - 1 downTo 0) {
-        days.add(Pair(lastDayOfPrevMonth - i, false))
-    }
-
-    // 현재 달의 날짜들
-    for (day in 1..lastDayOfMonth) {
-        days.add(Pair(day, true))
-    }
-
-    // 다음 달의 날짜들 (6주 * 7일 = 42개를 맞추기 위해)
-    val remainingDays = 42 - days.size
-    for (day in 1..remainingDays) {
-        days.add(Pair(day, false))
-    }
-
-    return days
-}
-
-
 @Preview(showBackground = true)
 @Composable
 fun RegisterDetailScreenPreview() {
     MaterialTheme {
         RegisterDetailScreen()
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun InfoModalOnlyPreview() {
-    InfoModal(
-        onDismiss    = {},
-        guestCount   = 100,
-        myGuestCount = 10
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun PriceInfoModalPreview() {
-    MaterialTheme {
-        PriceInfoModal(onDismiss = { })
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun CalendarModalPreview() {
-    MaterialTheme {
-        CalendarModal(
-            selectedDate = "2025-07-15",
-            onDateSelected = { },
-            onDismiss = { }
-        )
     }
 }
